Revert changes in strto(u)ll.c to avoid integer overflows
Found in oss-fuzz issue 26045.

Pick-to: 5.12 5.15

+++ src/3rdparty/freebsd/strtoll.c
@@ -127,7 +127,7 @@ qt_strtoll(const char * nptr, char **endptr, int base)
 noconv:
 		errno = EINVAL;
 	} else if (neg)
-		acc = (unsigned long long) -(long long)acc;
+		acc = -acc;
 	if (endptr != NULL)
                 *endptr = const_cast<char *>(any ? s - 1 : nptr);
 	return (acc);
